pyproject.toml
[tool.poetry]
name = "agri-energy-demo"
version = "0.1.0"
description = "MVP for Agri-Energy Intelligence Platform using RDF and Streamlit"
authors = ["Antigravity <antigravity@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
streamlit = "^1.32.0"
rdflib = "^7.0.0"
pyshacl = "^0.30.1"
pypdf = "^6.5.0"
requests = "^2.32.5"
beautifulsoup4 = "^4.14.3"
openai = "^2.14.0"
google-generativeai = "^0.8.6"
python-dotenv = "^1.2.1"
pymupdf4llm = "^0.2.7"
markdownify = "^1.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
